Add ES Dev Node shutdown
WAS - 65% : 458/699
NOW - 66% : 481/724

+++ src/main/java/org/devconferences/elastic/DeveloppementESNode.java
@@ -22,6 +22,8 @@ public class DeveloppementESNode {
     public static final String ES_LOCAL_DATA = "tmp/es-local-data";
     public static String elasticPort = "9200";
     private static final Logger LOGGER = LoggerFactory.getLogger(DeveloppementESNode.class);
+    static Node esNode = null;
+    static String portNode = null;
 
     public static void createDevNode() {
         createDevNode(elasticPort);
@@ -32,27 +34,42 @@ public class DeveloppementESNode {
     }
 
     public static void createDevNode(String port) {
-        LOGGER.info("Creating dev ES node on port " + port + "...");
-        Path localDevDataDirectory = Paths.get(ES_LOCAL_DATA);
-        try {
-            FileUtils.deleteDirectory(localDevDataDirectory.toFile());
-        } catch (IOException e) {
-            throw new RuntimeException(e);
+        if(esNode == null) {
+            LOGGER.info("Creating dev ES node on port " + port + "...");
+            Path localDevDataDirectory = Paths.get(ES_LOCAL_DATA);
+            try {
+                FileUtils.deleteDirectory(localDevDataDirectory.toFile());
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+
+            Settings settings = ImmutableSettings.builder()
+                    .put("http.port", port)
+                    .put("network.host", "localhost")
+                    .put("path.data", ES_LOCAL_DATA)
+                    .build();
+
+            esNode = NodeBuilder.nodeBuilder()
+                    .local(true)
+                    .data(true)
+                    .clusterName("elasticSearch" + UUID.randomUUID())
+                    .settings(settings)
+                    .build();
+            esNode.start();
+            portNode = port;
+        } else {
+            LOGGER.warn("ES Dev Node already launched on port " + portNode);
         }
-
-        Settings settings = ImmutableSettings.builder()
-                .put("http.port", port)
-                .put("network.host", "localhost")
-                .put("path.data", ES_LOCAL_DATA)
-                .build();
-
-        Node node = NodeBuilder.nodeBuilder()
-                .local(true)
-                .data(true)
-                .clusterName("elasticSearch" + UUID.randomUUID())
-                .settings(settings)
-                .build();
-        node.start();
     }
 
+    public static void deleteDevNode() {
+        if(esNode != null) {
+            LOGGER.info("Deleting dev ES node on port " + portNode + "...");
+            esNode.close();
+            esNode = null;
+            portNode = null;
+        } else {
+            LOGGER.warn("ES Dev Node not exist");
+        }
+    }
 }
